Normalize Text

Construct a standardized string from values passed to ...

Usage

normalize(...)

Arguments

...

Any number of vectors containing atomic elements. Each vector is normalized as a paragraph.

Elements are coerced to character values.

NA values and empty strings are discarded.

Multi-line strings are supported and encouraged. Blank lines are interpreted (two or more newline characters) as paragraph separators.

Value

A character string, possibly empty.

Details

Input text can written in a variety of ways using single-line and multi-line strings. Values passed to ... are normalized (to ensure their consistency) and collapsed to a single character string using the standard paragraph separator. The latter is defined as two newline characters ("\n\n").

NA values and empty strings are discarded before reducing ... to a character string.

Whitespaces (tabs, newlines, and repeated spaces) characters are replaced by a single space. Paragraph separators are preserved.

Leading or trailing whitespaces are stripped.